// modules/transfer/interface/table.enum.ts
export enum TableOption {
  CREATE = 'create_option',
  READ = 'read_option',
  UPDATE = 'update_option',
  MODIFY = 'modify_option',
  DELETE = 'delete_option',
  DELETE_FAVORITE = 'delete_favorite_option',
  TRANSFER = 'transfer_option',
  PAYMENT = 'payment_option',
  CHANGE_FAVORITE = 'change_favorite_option',
  ADD_FAVORITE = 'add_favorite_option',
}

export enum ESides {
  RIGHT = 'right',
  LEFT = 'left',
}