// modules/loan/modules/third-party-loans/enum/table-third-party-loans.enum.ts
export enum TableThirdPartyLoansOption {
  CREATE = 'create_option',
  READ = 'read_option',
  UPDATE = 'update_option',
  DELETE = 'delete_option',
  DELETE_FAVORITE = 'delete_favorite_option',
  CONSULT = 'consult_option',
  VIEW_DETAIL = 'viewDetail_option',
  MODIFY = 'modify_option',
  WATCH = 'option.watch',
  OPTION_DELETE = 'option.delete',
  CHANGE = 'option.change',
  SEND = 'option.send',
  WATCH_SIGNATURE = 'option.watch_signature',
  SEND_BACK = 'option.send_back',
  AUTHORIZED = 'option.authorized',
  PROCESS = 'option.process',
  TO_AUTHORIZE_SEND_BACK = 'toAuthorizeSendBack',
  AUTHORIZED_SEND_BACK = 'authorizedSendBack',
  TO_AUTHORIZE_VIEW_SIGNATURE = 'toAuthorizeViewSignature',
  AUTHORIZED_VIEW_SIGNATURE = 'authorizedViewSignature',
  PAYMENT = 'payment_option',
  ACCOUNT_STATEMENT = 'account-statement_option',
  NOT_ALLOWED = 'not_allowed'
}